顶顶顶

+++ model/jpa/src/main/java/org/keycloak/models/jpa/JpaRealmProvider.java
@@ -45,6 +45,8 @@ import javax.persistence.TypedQuery;
 
 import java.util.*;
 import java.util.stream.Collectors;
+import org.keycloak.models.ModelException;
+
 
 /**
  * @author <a href="mailto:bill@burkecentral.com">Bill Burke</a>
@@ -119,7 +121,6 @@ public class JpaRealmProvider implements RealmProvider {
             RealmModel realm = session.realms().getRealm(id);
             if (realm != null) realms.add(realm);
             em.flush();
-            em.clear();
         }
         return realms;
     }
@@ -293,6 +294,67 @@ public class JpaRealmProvider implements RealmProvider {
         }
         return list;
     }
+    
+    @Override
+    public Set<RoleModel> getRealmRoles(RealmModel realm, Integer first, Integer max) {
+        TypedQuery<RoleEntity> query = em.createNamedQuery("getRealmRoles", RoleEntity.class);
+        query.setParameter("realm", realm.getId());
+        
+        return getRoles(query, realm, first, max);
+    }
+
+    @Override
+    public Set<RoleModel> getClientRoles(RealmModel realm, ClientModel client, Integer first, Integer max) {
+        TypedQuery<RoleEntity> query = em.createNamedQuery("getClientRoles", RoleEntity.class);
+        query.setParameter("client", client.getId());
+        
+        return getRoles(query, realm, first, max);
+    }
+    
+    protected Set<RoleModel> getRoles(TypedQuery<RoleEntity> query, RealmModel realm, Integer first, Integer max) {
+        if(Objects.nonNull(first) && Objects.nonNull(max)
+                && first >= 0 && max >= 0) {
+            query= query.setFirstResult(first).setMaxResults(max);
+        }
+
+        List<RoleEntity> results = query.getResultList();
+        
+        return results.stream()
+                .map(role -> new RoleAdapter(session, realm, em, role))
+                .collect(Collectors.collectingAndThen(
+                        Collectors.toCollection(LinkedHashSet::new), Collections::unmodifiableSet));
+    }
+    
+    @Override
+    public Set<RoleModel> searchForClientRoles(RealmModel realm, ClientModel client, String search, Integer first, Integer max) {
+        TypedQuery<RoleEntity> query = em.createNamedQuery("searchForClientRoles", RoleEntity.class);
+        query.setParameter("client", client.getId());
+        return searchForRoles(query, realm, search, first, max);
+    }
+    
+    @Override
+    public Set<RoleModel> searchForRoles(RealmModel realm, String search, Integer first, Integer max) {
+        TypedQuery<RoleEntity> query = em.createNamedQuery("searchForRealmRoles", RoleEntity.class);
+        query.setParameter("realm", realm.getId());
+        
+        return searchForRoles(query, realm, search, first, max);
+    }
+    
+    protected Set<RoleModel> searchForRoles(TypedQuery<RoleEntity> query, RealmModel realm, String search, Integer first, Integer max) {
+
+        query.setParameter("search", "%" + search.trim().toLowerCase() + "%");
+        if(Objects.nonNull(first) && Objects.nonNull(max)
+                && first >= 0 && max >= 0) {
+            query= query.setFirstResult(first).setMaxResults(max);
+        }
+        
+        List<RoleEntity> results = query.getResultList();
+        
+        return results.stream()
+                .map(role -> new RoleAdapter(session, realm, em, role))
+                .collect(Collectors.collectingAndThen(
+                        Collectors.toSet(), Collections::unmodifiableSet));
+    }
 
     @Override
     public boolean removeRole(RealmModel realm, RoleModel role) {
@@ -302,6 +364,10 @@ public class JpaRealmProvider implements RealmProvider {
             container.removeDefaultRoles(role.getName());
         }
         RoleEntity roleEntity = em.getReference(RoleEntity.class, role.getId());
+        if (roleEntity == null || !roleEntity.getRealmId().equals(realm.getId())) {
+            // Throw model exception to ensure transaction rollback and revert previous operations (removing default roles) as well
+            throw new ModelException("Role not found or trying to remove role from incorrect realm");
+        }
         String compositeRoleTable = JpaUtils.getTableNameForNativeQuery("COMPOSITE_ROLE", em);
         em.createNativeQuery("delete from " + compositeRoleTable + " where CHILD_ROLE = :role").setParameter("role", roleEntity).executeUpdate();
         realm.getClients().forEach(c -> c.deleteScopeMapping(role));
@@ -372,15 +438,16 @@ public class JpaRealmProvider implements RealmProvider {
 
     @Override
     public Long getGroupsCount(RealmModel realm, Boolean onlyTopGroups) {
-        String query = "getGroupCount";
         if(Objects.equals(onlyTopGroups, Boolean.TRUE)) {
-            query = "getTopLevelGroupCount";
+            return em.createNamedQuery("getTopLevelGroupCount", Long.class)
+                    .setParameter("realm", realm.getId())
+                    .setParameter("parent", GroupEntity.TOP_PARENT_ID)
+                    .getSingleResult();
+        } else {
+            return em.createNamedQuery("getGroupCount", Long.class)
+                    .setParameter("realm", realm.getId())
+                    .getSingleResult();
         }
-        Long count = em.createNamedQuery(query, Long.class)
-                .setParameter("realm", realm.getId())
-                .getSingleResult();
-
-        return count;
     }
 
     @Override
@@ -419,7 +486,7 @@ public class JpaRealmProvider implements RealmProvider {
         RealmEntity ref = em.getReference(RealmEntity.class, realm.getId());
 
         return ref.getGroups().stream()
-                .filter(g -> g.getParent() == null)
+                .filter(g -> GroupEntity.TOP_PARENT_ID.equals(g.getParentId()))
                 .map(g -> session.realms().getGroupById(g.getId(), realm))
                 .sorted(Comparator.comparing(GroupModel::getName))
                 .collect(Collectors.collectingAndThen(
@@ -430,6 +497,7 @@ public class JpaRealmProvider implements RealmProvider {
     public List<GroupModel> getTopLevelGroups(RealmModel realm, Integer first, Integer max) {
         List<String> groupIds =  em.createNamedQuery("getTopLevelGroupIds", String.class)
                 .setParameter("realm", realm.getId())
+                .setParameter("parent", GroupEntity.TOP_PARENT_ID)
                 .setFirstResult(first)
                     .setMaxResults(max)
                     .getResultList();
@@ -440,9 +508,7 @@ public class JpaRealmProvider implements RealmProvider {
                 list.add(group);
             }
         }
-
-        list.sort(Comparator.comparing(GroupModel::getName));
-
+        // no need to sort, it's sorted at database level
         return Collections.unmodifiableList(list);
     }
 
@@ -492,19 +558,19 @@ public class JpaRealmProvider implements RealmProvider {
     }
 
     @Override
-    public GroupModel createGroup(RealmModel realm, String name) {
-        String id = KeycloakModelUtils.generateId();
-        return createGroup(realm, id, name);
-    }
-
-    @Override
-    public GroupModel createGroup(RealmModel realm, String id, String name) {
-        if (id == null) id = KeycloakModelUtils.generateId();
+    public GroupModel createGroup(RealmModel realm, String id, String name, GroupModel toParent) {
+        if (id == null) {
+            id = KeycloakModelUtils.generateId();
+        } else if (GroupEntity.TOP_PARENT_ID.equals(id)) {
+            // maybe it's impossible but better ensure this doesn't happen
+            throw new ModelException("The ID of the new group is equals to the tag used for top level groups");
+        }
         GroupEntity groupEntity = new GroupEntity();
         groupEntity.setId(id);
         groupEntity.setName(name);
         RealmEntity realmEntity = em.getReference(RealmEntity.class, realm.getId());
         groupEntity.setRealm(realmEntity);
+        groupEntity.setParentId(toParent == null? GroupEntity.TOP_PARENT_ID : toParent.getId());
         em.persist(groupEntity);
         em.flush();
         realmEntity.getGroups().add(groupEntity);
@@ -572,6 +638,20 @@ public class JpaRealmProvider implements RealmProvider {
     @Override
     public List<ClientModel> getClients(RealmModel realm) {
         return this.getClients(realm, null, null);
+    }
+
+    @Override
+    public List<ClientModel> getAlwaysDisplayInConsoleClients(RealmModel realm) {
+        TypedQuery<String> query = em.createNamedQuery("getAlwaysDisplayInConsoleClients", String.class);
+        query.setParameter("realm", realm.getId());
+        List<String> clients = query.getResultList();
+        if (clients.isEmpty()) return Collections.EMPTY_LIST;
+        List<ClientModel> list = new LinkedList<>();
+        for (String id : clients) {
+            ClientModel client = session.realms().getClientById(id, realm);
+            if (client != null) list.add(client);
+        }
+        return Collections.unmodifiableList(list);
     }
 
     @Override
@@ -710,20 +790,18 @@ public class JpaRealmProvider implements RealmProvider {
     @Override
     public ClientInitialAccessModel getClientInitialAccessModel(RealmModel realm, String id) {
         ClientInitialAccessEntity entity = em.find(ClientInitialAccessEntity.class, id);
-        if (entity == null) {
-            return null;
-        } else {
-            return entityToModel(entity);
-        }
+        if (entity == null) return null;
+        if (!entity.getRealm().getId().equals(realm.getId())) return null;
+        return entityToModel(entity);
     }
 
     @Override
     public void removeClientInitialAccessModel(RealmModel realm, String id) {
         ClientInitialAccessEntity entity = em.find(ClientInitialAccessEntity.class, id, LockModeType.PESSIMISTIC_WRITE);
-        if (entity != null) {
-            em.remove(entity);
-            em.flush();
-        }
+        if (entity == null) return;
+        if (!entity.getRealm().getId().equals(realm.getId())) return;
+        em.remove(entity);
+        em.flush();
     }
 
     @Override
@@ -764,4 +842,5 @@ public class JpaRealmProvider implements RealmProvider {
         model.setTimestamp(entity.getTimestamp());
         return model;
     }
+
 }
